requirements.txt
fitbit==0.3.1
supabase==1.0.2
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
cherrypy==18.8.0
requests==2.31.0